python/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
asyncpg==0.28.0
redis==5.0.0
camelot-py==0.9.0
PyMuPDF==1.23.5
pytesseract==0.3.10
Pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
spacy==3.7.2
sentence-transformers==2.7.0
ollama==0.1.9
opencv-python==4.8.1.78
psycopg2-binary>=2.9.0
python-multipart==0.0.6
git+https://github.com/openai/CLIP.git@main#egg=clip